ssh connection check and removal confirmation

+++ webapp/public/javascripts/angular/table/table.js
@@ -14,13 +14,13 @@ angular.module('terrama2.table', ['terrama2'])
         remove: '&?',
         extra: '=?extra'
       },
-      
+
       controller: function($scope, $http, i18n) {
         $scope.i18n = i18n;
         $scope.searchInput = '';
         $scope.selected = {};
         $scope.emptyMessage = 'No ' + ($scope.context || 'data') + ' found';
-        
+
         // defines display fields in table
         $scope.displayFields = [];
         // fields identifiers
@@ -28,35 +28,46 @@ angular.module('terrama2.table', ['terrama2'])
 
         $scope.extra = $scope.extra ? $scope.extra : {};
 
+        $scope.objectToRemove = null;
+
         $scope.setModelSelected = function(modelSelected) {
           $scope.selected = modelSelected;
         }
 
+        $scope.confirmRemoval = function(object) {
+          $scope.objectToRemove = object;
+
+          $('#myModal').modal();
+        };
+
         // remove function
-        $scope.removeOperation = function(object) {
-          // todo: open model: confirmation
+        $scope.removeOperation = function() {
+          if($scope.objectToRemove !== null) {
+            var object = $scope.objectToRemove;
+            $scope.objectToRemove = null;
 
-          // callback
-          var callback = $scope.extra.removeOperationCallback;
-          $http({
-            method: 'DELETE',
-            url: $scope.remove({object: object})
-          }).success(function(response) {
-            $scope.model.forEach(function(element, index, arr) {
-              if (element.id == object.id)
-                arr.splice(index, 1);
+            // callback
+            var callback = $scope.extra.removeOperationCallback;
+            $http({
+              method: 'DELETE',
+              url: $scope.remove({object: object})
+            }).success(function(response) {
+              $scope.model.forEach(function(element, index, arr) {
+                if (element.id == object.id)
+                  arr.splice(index, 1);
 
+                if ($scope.isFunction(callback))
+                  callback(null, response);
+              });
+
+            }).error(function(err) {
               if ($scope.isFunction(callback))
-                callback(null, response);
+                callback(err);
+              console.log(err);
+            }).finally(function() {
+              // $scope.selected = {};
             });
-
-          }).error(function(err) {
-            if ($scope.isFunction(callback))
-              callback(err);
-            console.log(err);
-          }).finally(function() {
-            // $scope.selected = {};
-          });
+          }
         };
 
         if (!$scope.iconProperties)
@@ -83,7 +94,7 @@ angular.module('terrama2.table', ['terrama2'])
         $scope.isFunction = function(target) {
           return angular.isFunction(target);
         };
-        
+
         $scope.capitalizeIt = function(str) {
           return str.charAt(0).toUpperCase() + str.slice(1);
         }
